Add cluster name validator in the new model
Move unit test from test_commands to test_cluster
