+++ cli/src/pcluster/models/cluster.py
@@ -14,6 +14,7 @@
 #
 import json
 import logging
+import re
 import time
 from enum import Enum
 from typing import List
@@ -27,7 +28,7 @@ from common.aws.aws_resources import InstanceInfo, StackInfo
 from common.boto3.common import AWSClientError
 from pcluster.cli_commands.compute_fleet_status_manager import ComputeFleetStatus, ComputeFleetStatusManager
 from pcluster.config.config_patch import ConfigPatch
-from pcluster.constants import OS_MAPPING, PCLUSTER_STACK_PREFIX
+from pcluster.constants import OS_MAPPING, PCLUSTER_NAME_MAX_LENGTH, PCLUSTER_NAME_REGEX, PCLUSTER_STACK_PREFIX
 from pcluster.models.cluster_config import BaseClusterConfig, ClusterBucket, SlurmScheduling, Tag
 from pcluster.schemas.cluster_schema import ClusterSchema
 from pcluster.templates.cdk_builder import CDKTemplateBuilder
@@ -279,11 +280,13 @@ class Cluster:
             try:
                 # syntactic validation
                 self.config = ClusterSchema().load(self.source_config)
-                if not suppress_validators:
-                    # semantic validation
-                    LOGGER.info("Validating cluster configuration...")
-                    validation_failures = self.config.validate()
 
+                # semantic validation
+                if not suppress_validators:
+                    validation_failures = self._validate_cluster_name()
+
+                    LOGGER.info("Validating cluster configuration...")
+                    validation_failures += self.config.validate()
                     for failure in validation_failures:
                         if failure.level.value >= FailureLevel(validation_failure_level).value:
                             # Raise the exception if there is a failure with a level greater than the specified one
@@ -331,6 +334,18 @@ class Cluster:
                 # Cleanup S3 artifacts if stack is not created yet
                 self.bucket.delete()
             raise ClusterActionError(f"Cluster creation failed.\n{e}")
+
+    def _validate_cluster_name(self):
+        validation_failures = []
+        if not re.match(PCLUSTER_NAME_REGEX % (PCLUSTER_NAME_MAX_LENGTH - 1), self.name):
+            message = (
+                "Error: The cluster name can contain only alphanumeric characters (case-sensitive) and hyphens. "
+                "It must start with an alphabetic character and can't be longer "
+                f"than {PCLUSTER_NAME_MAX_LENGTH} characters."
+            )
+            LOGGER.error(message)
+            validation_failures.append(ValidationResult(str(message), FailureLevel.ERROR))
+        return validation_failures
 
     def _setup_cluster_bucket(self) -> ClusterBucket:
         """
